sched/env: add tg_envc in task_group_s to avoid some loops in code
this commit optimize the env api speed

+++ sched/environ/env_getenv.c
@@ -61,7 +61,7 @@ FAR char *getenv(FAR const char *name)
   FAR struct tcb_s *rtcb;
   FAR struct task_group_s *group;
   FAR char *pvalue = NULL;
-  int ret = OK;
+  ssize_t ret = OK;
 
   /* Verify that a string was passed */
